Refactor vision tests
