+++ src/main/java/frc/robot/RobotContainer.java
@@ -23,7 +23,7 @@ import frc.robot.commands.test.TestDriveVelocity;
 import frc.robot.commands.test.TestHeadingPID;
 import frc.robot.commands.test.TestSteerAngle;
 import frc.robot.commands.vision.AngleAlignTest;
-import frc.robot.commands.vision.TestVision;
+import frc.robot.commands.vision.TestVisionDistance;
 import frc.robot.commands.vision.VisionReturnTest;
 import frc.robot.constants.VisionConstants;
 import frc.robot.constants.swerve.DriveConstants;
@@ -164,8 +164,8 @@ public class RobotContainer {
     m_testTab.add("Heading PID", new TestHeadingPID(m_drive, testEntry));
     m_testTab.add("Steer angle", new TestSteerAngle(m_drive, testEntry));
     m_testTab.add("Odometry Test", new OdometryTestCommand(m_drive, new Transform2d(new Translation2d(1,1), new Rotation2d(Math.PI))));
-    m_testTab.add("Test vision (forward)", new TestVision(0.2, m_drive, m_vision));
-    m_testTab.add("Test vision (backward)", new TestVision(-0.2, m_drive, m_vision));
+    m_testTab.add("Test vision (forward)", new TestVisionDistance(0.2, m_drive, m_vision));
+    m_testTab.add("Test vision (backward)", new TestVisionDistance(-0.2, m_drive, m_vision));
     m_testTab.add("Vision return test (forward)", new VisionReturnTest(0.2, 3, m_drive, m_vision));
     m_testTab.add("Vision return test (backward)", new VisionReturnTest(-+ 0.2, 3, m_drive, m_vision));
     m_testTab.add("Align to 0 degrees", new AngleAlignTest(0, m_drive, m_vision));
