+++ src/test/java/frc/robot/util/FieldLayoutTest.java
@@ -1,12 +1,16 @@
 package frc.robot.util;
 
 import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.fail;
+
+import java.io.IOException;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import edu.wpi.first.apriltag.AprilTagFieldLayout;
+import edu.wpi.first.apriltag.AprilTagFields;
 import edu.wpi.first.wpilibj.shuffleboard.Shuffleboard;
 import frc.robot.constants.VisionConstants;
 
@@ -18,7 +22,7 @@ import frc.robot.constants.VisionConstants;
  * To disable a test, annotate with Disabled
  */
 public class FieldLayoutTest {
-  
+
   private AprilTagFieldLayout fieldLayout;
 
   @BeforeEach
@@ -51,5 +55,20 @@ public class FieldLayoutTest {
   @Test
   public void test2Above1() {
     assertTrue(fieldLayout.getTagPose(2).get().getY()>fieldLayout.getTagPose(1).get().getY());
+  }
+
+  /**
+   * Checks that the 2023ChargedUp AprilTag field layout provided by FIRST matches the one in VisionConstants.kAprilTags
+   */
+  @Test
+  public void testAprilTagPoses() {
+    try {
+      AprilTagFieldLayout aprilTagFieldLayout = AprilTagFields.k2023ChargedUp.loadAprilTagLayoutField();
+      for (int i = 0; i < 8; i++) {
+        assertTrue(aprilTagFieldLayout.getTagPose(i + 1).get().equals(VisionConstants.kAprilTags.get(i).pose), "AprilTag " + Integer.toString(i + 1) + " doesn't match");
+      }
+    } catch (IOException e) {
+      fail("Could not load k2023ChargedUp file from WPILib, check that GradleRIO version >= 2023.2.1");
+    }
   }
 }
